more more cleaning up qe_wrapper [wip]
