+++ simespresso/io_process.py
@@ -6,6 +6,8 @@ import os
 import subprocess
 import logging
 
+from simespresso.io_QE.data_handler import  qe_data_handler
+
 class QeProcess(object):
     """ Class runs the lammps/liggghts program
     Parameters
@@ -25,6 +27,8 @@ class QeProcess(object):
         self._returncode = 0
         self._stderr = ""
         self._stdout = ""
+        self._datahandler = qe_data_handler()
+
         if log_directory:
             self._log = os.path.join(log_directory, 'log.qe')
         else:
@@ -41,7 +45,12 @@ class QeProcess(object):
                 msg += " stdout/err: " + self._stdout + " " + self._stderr
             raise RuntimeError(msg)
 
-    def run(self, commands):
+    def run(self,
+                input_data_file=input_data_filename,
+                output_data_file=output_data_filename,
+                BC=_combine(self.BC, self.BC_extension),
+                CM=_combine(self.CM, self.CM_extension),
+                SP=_combine(self.SP, self.SP_extension)):
         """Run engine with a set of commands
         Parameters
         ----------
@@ -54,12 +63,9 @@ class QeProcess(object):
         """
         logging.debug('starting qe engine')
 
-
-    pwname = self.datahandler.input_pwname
-        self.datahandler.write_espresso_input_file(pwname)
-    path = self.datahandler.path_to_espresso
-        logging.debug('path to espresso:'+path)
-        if not which(self.path_to_espresso):
+        self._datahandler.write_espresso_input_file(input_data_filename
+        logging.debug('path to espresso:'+self._qe_executable)
+        if not which(self._qe_executable):
             logging.debug('no path to espresso')
             raise ValueError(
                 'espresso command not found (looking for '
@@ -100,3 +106,20 @@ class QeProcess(object):
                 msg += "stdout: \'{}\n\'".format(self._stdout)
             raise RuntimeError(msg)
 '''
+
+
+def which(name):
+    found = 0
+    for path in os.getenv("PATH").split(os.path.pathsep):
+        full_path = path + os.sep + name
+        if os.path.exists(full_path):
+            """
+            if os.stat(full_path).st_mode & stat.S_IXUSR:
+                found = 1
+                print(full_path)
+            """
+            found = 1
+            print(full_path)
+    # Return a UNIX-style exit code so it can be checked by calling scripts.
+    # Programming shortcut to toggle the value of found: 1 => 0, 0 => 1.
+    sys.exit(1 - found)
